+++ bin/lib/templates/LoopBaseTemplate.js
