+++ bin/lib/templates/LoopTemplate.js
