+++ bin/lib/templates/TraceTemplate.js
