+++ bin/lib/templates/index.js
