+++ lib/templates/LoopBaseTemplate.ts
