+++ lib/templates/LoopTemplate.ts
