+++ lib/templates/TraceTemplate.ts
