+++ lib/templates/index.ts
